Renamed batch folder to batch; removed ununsed crates from Cargo.toml; refactored batch example to use common example pattern
