+++ aws/sdk/examples/batch/Cargo.toml
@@ -1,14 +1,14 @@
 [package]
-name = "batch-helloworld"
+name = "batch-code-examples"
 version = "0.1.0"
-authors = ["Alistair McLean <mclean@amazon.com>"]
+authors = ["Alistair McLean <mclean@amazon.com>", "Doug Schwartz <dougsch@amazon.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 batch = { package = "aws-sdk-batch", path = "../../build/aws-sdk/batch" }
+aws-types = { path = "../../build/aws-sdk/aws-types" }
 tokio = { version = "1", features = ["full"] }
-# used only to enable basic logging:
-env_logger = "0.8.2"
+structopt = { version = "0.3", default-features = false }
 tracing-subscriber = "0.2.18"
